New: common base class for working with VMs
